Add ITunes Title (#44)
The itunes: namespace and Apple Podcasts support a `<itunes:title>` element. This
is mostly redundant with `<title>`, but is intended to be used to strip out episode
and season numbers when using `<itunes:episode>` and `<itunes:season>`. This adds
support for `<itunes:title>` on RSS items through the `item.itunes_item` method.

Reference for including itunes:title: https://help.omnystudio.com/en/articles/3219724-setting-a-separate-itunes-title-and-title-in-your-rss-feed-for-episode-titles

Sample feed including an itunes:title: https://raw.githubusercontent.com/bcomnes/jsonfeed-to-rss/master/reference/podcast.xml

Co-authored-by: Ryan Brunner <ryan@supercast.com>