Use `cgi/escape` instead of `cgi/util` (#62)
### Background
A warning is currently emitted in CI when running with ruby-head (Ruby
3.5-dev).
This warning occurs because `CGI::Util` will be removed in Ruby 3.5.
- https://bugs.ruby-lang.org/issues/21258
- https://github.com/ruby/ruby/pull/13275

```
/home/runner/.rubies/ruby-head/lib/ruby/gems/3.5.0+2/gems/rss-0.3.2/lib/rss/rss.rb:4: warning: CGI::Util is removed from Ruby 3.5. Please use cgi/escape instead for CGI.escape and CGI.unescape features.
If you are using CGI.parse, please install and use the cgi gem instead.
/home/runner/work/rss/rss/tmp/test/test-dublincore.rb:2: warning: CGI library is removed from Ruby 3.5. Please use cgi/escape instead for CGI.escape and CGI.unescape features.
If you need to use the full features of CGI library, Please install cgi gem.
```

https://github.com/ruby/rss/actions/runs/16902042489/job/47883073913#step:6:8

### Details
This Pull Request fixes the warning by replacing the use of `cgi/util`
with `cgi/escape`.
The `cgi/escape` is available since Ruby 2.3.

### Additional Information
Similar fixes have been made in other libraries, such as the rdoc gem.
- https://github.com/ruby/rdoc/pull/1366
- https://github.com/ruby/rdoc/pull/1378
-
https://github.com/ruby/net-http/commit/d8fd39c589279b1aaec85a7c8de9b3e199c72efe